dom/base/test/test_pluginAudioNotification.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for audio controller in windows</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="plugin.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<pre id="test">
</pre>
<iframe></iframe>

<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

var expectedNotification = null;
var iframe = null;

var observer = {
  observe: function(subject, topic, data) {
    is(topic, "audio-playback", "audio-playback received");
    is(data, expectedNotification, "This is the right notification");
    SimpleTest.executeSoon(runTest);
  }
};

var observerService = SpecialPowers.Cc["@mozilla.org/observer-service;1"]
                                   .getService(SpecialPowers.Ci.nsIObserverService);

var tests = [
  function() {
    iframe = document.querySelector("iframe");
    observerService.addObserver(observer, "audio-playback", false);
    ok(true, "Observer set");
    runTest();
  },

  function() {
    expectedNotification = 'active';
    iframe.src = "file_pluginAudio.html";
  },

  function() {
    info("=== Mute plugin ===");
    ok(!iframe.contentWindow.pluginMuted(), "Plugin should not be muted");
    iframe.contentWindow.toggleMuteState(true);
    ok(iframe.contentWindow.pluginMuted(), "Plugin should be muted");
    expectedNotification = 'inactive-nonaudible';
  },

  function() {
    info("=== unmute plugin ==");
    ok(iframe.contentWindow.pluginMuted(), "Plugin should be muted");
    iframe.contentWindow.toggleMuteState(false);
    ok(!iframe.contentWindow.pluginMuted(), "Plugin should not be muted");
    expectedNotification = 'active';
  },

  function() {
    info("=== stop audio ==");
    expectedNotification = 'inactive-pause';
    iframe.contentWindow.stopAudio();
  },

  function() {
    observerService.removeObserver(observer, "audio-playback");
    ok(true, "Observer removed");
    runTest();
  }
];

function runTest() {
  if (!tests.length) {
    SimpleTest.finish();
    return;
  }

  var test = tests.shift();
  test();
}

onload = runTest;

</script>
</body>
</html>
